rename torch to data module
this is a pytorch application so clearer to name it data
